Following Changes incorporated
1. Added New POM File - Compatible for relesing to Maven Central
2. Moved Test cases to the same bundle (src/test/java)
3. Corrected Package Name Typo
4. Extracted all annotations from the BundleTestSuite because of
compilation problem in Maven Compiler Plugin
5. Source Folder changed from src to src/main/java
6. Modified BundleTestSuite javadoc to eliminate javadoc plugin
compilation problem
7. Removed releng dependency
8. Modified README.md to incorporate future Maven Central Dependency
