--- com.codeaffine.osgi.testsuite.releng/feature/build.properties
@@ -1,12 +1,0 @@
-###############################################################################
-# Copyright (c) 2012, 2013, 2014 Rüdiger Herrmann.
-# All rights reserved. This program and the accompanying materials
-# are made available under the terms of the Eclipse Public License v1.0
-# which accompanies this distribution, and is available at
-# http://www.eclipse.org/legal/epl-v10.html
-#
-# Contributors:
-#    Rüdiger Herrmann - initial API and implementation
-################################################################################
-
-bin.includes = feature.xml,license.html
