--- com.codeaffine.osgi.testsuite.releng/feature/feature.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2012, 2013 Rüdiger Herrmann.
-  All rights reserved. This program and the accompanying materials
-  are made available under the terms of the Eclipse Public License v1.0
-  which accompanies this distribution, and is available at
-  http://www.eclipse.org/legal/epl-v10.html
-
-  Contributors:
-     Rüdiger Herrmann - initial implementation
--->
-<feature
-      id="com.codeaffine.osgi.testsuite.feature"
-      label="Automated OSGi Test Suite"
-      version="1.2.0.qualifier"
-      provider-name="Codeaffine.com">
-
-   <description url="">
-      The OSGi Test Suite runs all JUnit tests in a given list of bundles.
-      For example:
-        @RunWith( BundleTestSuite.class )
-        @TestBundles( { &quot;org.example.bundle1&quot;, &quot;org.example.bundle2&quot; } )
-        public class AllTestSuite {
-        }
-   </description>
-
-   <copyright>
-      Copyright 2013, 2014 Rüdiger Herrmann
-   </copyright>
-
-   <license url="license.html">
-      
-   </license>
-
-   <plugin
-         id="com.codeaffine.osgi.testsuite"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-   <plugin
-         id="com.codeaffine.osgi.testsuite.source"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-</feature>
