--- com.codeaffine.osgi.testsuite.releng/feature/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>com.codeaffine.osgi.testsuite</groupId>
-    <artifactId>com.codeaffine.osgi.testsuite.parent</artifactId>
-    <version>1.2.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>com.codeaffine.osgi.testsuite.feature</artifactId>
-  <packaging>eclipse-feature</packaging>
-
-</project>
