--- com.codeaffine.osgi.testsuite.releng/pom.xml
@@ -1,166 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2012, 2013 Rüdiger Herrmann.
-  All rights reserved. This program and the accompanying materials
-  are made available under the terms of the Eclipse Public License v1.0
-  which accompanies this distribution, and is available at
-  http://www.eclipse.org/legal/epl-v10.html
-
-  Contributors:
-     Rüdiger Herrmann - initial implementation
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <name>Automated OSGi Test Suite</name>
-
-  <groupId>com.codeaffine.osgi.testsuite</groupId>
-  <artifactId>com.codeaffine.osgi.testsuite.parent</artifactId>
-  <version>1.2.0-SNAPSHOT</version>
-  <packaging>pom</packaging>
-
-  <properties>
-    <tycho-version>0.23.0</tycho-version>
-    <tycho-extras-version>0.23.0</tycho-extras-version>
-    <surefire-version>2.19</surefire-version>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <tycho.scmUrl>scm:git:https://github.com/rherrmann/osgi-testsuite.git</tycho.scmUrl>
-  </properties>
-
-  <prerequisites>
-    <maven>3.0</maven>
-  </prerequisites>
-
-  <modules>
-    <module>feature</module>
-    <module>repository</module>
-    <module>../com.codeaffine.osgi.testsuite</module>
-    <module>../com.codeaffine.osgi.testsuite.test</module>
-  </modules>
-
-  <pluginRepositories>
-    <pluginRepository>
-      <id>repo.eclipse.org</id>
-      <url>https://repo.eclipse.org/content/repositories/releases/</url>
-    </pluginRepository>
-  </pluginRepositories>
-
-  <build>
-    <plugins>
-
-      <plugin>
-         <groupId>org.eclipse.tycho</groupId>
-         <artifactId>target-platform-configuration</artifactId>
-         <version>${tycho-version}</version>
-         <configuration>
-            <target>
-               <artifact>
-                  <groupId>com.codeaffine.osgi.testsuite</groupId>
-                  <artifactId>com.codeaffine.osgi.testsuite.parent</artifactId>
-                  <version>1.2.0-SNAPSHOT</version>
-                  <classifier>equinox</classifier>
-               </artifact>
-            </target>
-         </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.eclipse.tycho</groupId>
-        <artifactId>tycho-maven-plugin</artifactId>
-        <version>${tycho-version}</version>
-        <extensions>true</extensions>
-      </plugin>
-
-      <plugin>
-        <groupId>org.eclipse.tycho</groupId>
-        <artifactId>tycho-packaging-plugin</artifactId>
-        <version>${tycho-version}</version>
-        <dependencies>
-          <dependency>
-            <groupId>org.eclipse.tycho.extras</groupId>
-            <artifactId>tycho-sourceref-jgit</artifactId>
-            <version>${tycho-extras-version}</version>
-          </dependency>
-        </dependencies>        
-        <configuration>
-          <format>yyyyMMdd-HHmm</format>
-          <sourceReferences>
-            <generate>true</generate>
-          </sourceReferences>
-        </configuration>
-      </plugin>
-
-      <plugin>
-        <groupId>org.eclipse.tycho</groupId>
-        <artifactId>tycho-source-plugin</artifactId>
-        <version>${tycho-version}</version>
-        <executions>
-          <execution>
-            <id>plugin-source</id>
-            <goals>
-              <goal>plugin-source</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-surefire-plugin</artifactId>
-        <version>${surefire-version}</version>
-        <dependencies>
-          <dependency>
-            <groupId>org.apache.maven.surefire</groupId>
-            <artifactId>surefire-junit47</artifactId>
-            <version>${surefire-version}</version>
-          </dependency>
-        </dependencies>
-        <executions>
-          <execution>
-            <id>test</id>
-            <phase>test</phase>
-            <configuration>
-              <testClassesDirectory>${project.build.outputDirectory}</testClassesDirectory>
-              <includes>
-                <include>**/*Test.java</include>
-              </includes>
-              <excludes>
-                <!-- override default excludes to exclude nothing -->
-              </excludes>
-              <runOrder>random</runOrder>
-            </configuration>
-            <goals>
-              <goal>test</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-      
-      <plugin>
-        <groupId>org.codehaus.mojo</groupId>
-        <artifactId>animal-sniffer-maven-plugin</artifactId>
-        <version>1.14</version>
-        <configuration>
-          <signature>
-            <groupId>org.codehaus.mojo.signature</groupId>
-            <artifactId>java15</artifactId>
-            <version>1.0</version>
-          </signature>
-        </configuration>
-        <executions>
-          <execution>
-            <id>ensure-java-1.5-class-library</id>
-            <phase>test</phase>
-            <goals>
-              <goal>check</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-
-    </plugins>
-  </build>
-
-</project>
