--- com.codeaffine.osgi.testsuite.releng/repository/category.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2012, 2013 Rüdiger Herrmann.
-  All rights reserved. This program and the accompanying materials
-  are made available under the terms of the Eclipse Public License v1.0
-  which accompanies this distribution, and is available at
-  http://www.eclipse.org/legal/epl-v10.html
-
-  Contributors:
-     Rüdiger Herrmann - initial implementation
--->
-<site>
-   <feature 
-       url="../features/com.codeaffine.osgi.testsuite.feature_1.2.0.qualifier.jar" 
-       id="com.codeaffine.osgi.testsuite.feature" 
-       version="1.2.0.qualifier">
-      <category name="com.codeaffine.osgi.testsuite.category"/>
-   </feature>
-   <category-def name="com.codeaffine.osgi.testsuite.category" label="Automated OSGi Test Suite"/>
-</site>
