--- com.codeaffine.osgi.testsuite.releng/repository/pom.xml
@@ -1,106 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  Copyright (C) 2012, 2013, 2014 Rüdiger Herrmann.
-  All rights reserved. This program and the accompanying materials
-  are made available under the terms of the Eclipse Public License v1.0
-  which accompanies this distribution, and is available at
-  http://www.eclipse.org/legal/epl-v10.html
-
-  Contributors:
-     Rüdiger Herrmann - initial implementation
--->
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>com.codeaffine.osgi.testsuite</groupId>
-    <artifactId>com.codeaffine.osgi.testsuite.parent</artifactId>
-    <version>1.2.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>com.codeaffine.osgi.testsuite.repository</artifactId>
-  <packaging>eclipse-repository</packaging>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.eclipse.tycho</groupId>
-        <artifactId>tycho-p2-repository-plugin</artifactId>
-        <version>${tycho-version}</version>
-        <configuration>
-          <repositoryName>OSGi Test Suite Software Repository</repositoryName>
-          <includeAllDependencies>false</includeAllDependencies>
-          <compress>true</compress>
-        </configuration>
-      </plugin>
-      
-      <plugin>
-        <artifactId>maven-resources-plugin</artifactId>
-        <version>2.6</version>
-        <executions>
-          <execution>
-            <id>copy-resources</id>
-            <phase>package</phase>
-            <goals>
-              <goal>copy-resources</goal>
-            </goals>
-            <configuration>
-              <encoding>UTF-8</encoding>
-              <outputDirectory>${basedir}/target/repository</outputDirectory>
-              <resources>          
-                <resource>
-                  <directory>rootfiles</directory>
-                  <filtering>false</filtering>
-                </resource>
-              </resources>              
-            </configuration>            
-          </execution>
-        </executions>
-      </plugin>
-      
-      <plugin>
-        <groupId>org.eclipse.tycho.extras</groupId>
-        <artifactId>tycho-eclipserun-plugin</artifactId>
-        <version>${tycho-extras-version}</version>
-        <executions>
-          <execution>
-            <id>add-repo-properties</id>
-            <goals>
-              <goal>eclipse-run</goal>
-            </goals>
-            <phase>package</phase>
-            <configuration>
-              <appArgLine>-application org.eclipse.wtp.releng.tools.addRepoProperties</appArgLine>
-              <argLine>
-                -DartifactRepoDirectory=${project.build.directory}/repository
-                -Dp2StatsURI=http://www.codeaffine.com/p2-stats
-                -DstatsTrackedArtifacts=com.codeaffine.osgi.testsuite.feature
-                -DstatsArtifactsSuffix=-${project.version}
-              </argLine>
-              <dependencies>
-                <dependency>
-                  <artifactId>org.eclipse.wtp.releng.tools.feature</artifactId>
-                  <type>eclipse-feature</type>
-                </dependency>
-              </dependencies>
-              <repositories>
-                <repository>
-                  <url>http://download.eclipse.org/webtools/releng/repository</url>
-                  <layout>p2</layout>
-                </repository>
-                <repository>
-                  <url>http://download.eclipse.org/releases/kepler</url>
-                  <layout>p2</layout>
-                </repository>
-              </repositories>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-            
-    </plugins>
-  </build>
-
-</project>
