--- com.codeaffine.osgi.testsuite.releng/repository/rootfiles/index.html
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
-    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <title>Automated OSGi Test Suite Repository</title>
-    <style type="text/css">
-    <!--
-      * { font-family: sans-serif; }
-    -->
-    </style>
-  </head>
-  <body>
-    <h1>Automated OSGi Test Suite Repository</h1>
-    <p>
-      This is the software repository that contains the latest stable build for the 
-      <a href="https://github.com/rherrmann/osgi-testsuite">Automated OSGi Test Suite</a>.
-    </p>
-  </body>
-</html>
