--- com.codeaffine.osgi.testsuite/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,0 @@
-#Fri Jan 06 22:50:00 CET 2012
-eclipse.preferences.version=1
-encoding/<project>=UTF-8
