+++ com.codeaffine.osgi.testsuite/META-INF/MANIFEST.MF
@@ -12,5 +12,5 @@ Import-Package: org.junit;version="[4.8.0,5.0.0)",
  org.junit.runners.model;version="[4.8.0,5.0.0)",
  org.osgi.framework;version="[1.3.0,2.0.0)",
  org.osgi.framework.wiring;version="[1.0.0,2.0.0)"
-Export-Package: com.codeaffine.osgi.testuite;version="1.2.0"
+Export-Package: com.codeaffine.osgi.testsuite;version="1.2.0"
 Bundle-ActivationPolicy: lazy
